refactor: move syncing functionality to sync actions only
